// src/app/shared/Models/UserAdmin.ts
import { Rol } from "./Rol";

export class UserAdmin {

    id?: number;
    username?: string;
    passAccess?: string;
    idStatus?: number;
    attempt?: number;
    idTypeUser?: number;
    idPerson?: number;
    idPerfil?: number;
    startIdUser?: number;
    startDate?: number;
    updateIdUser?: number;
    updateDate?: number;
    boolDelete?: boolean;

}